requirements.txt
blinker==1.6.2
click==8.1.3
distlib==0.3.6
filelock==3.12.2
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
paho-mqtt==1.6.1
pexif==0.15
Pillow==9.4.0
platformdirs==3.8.1
virtualenv==20.23.1
waitress==2.1.2
Werkzeug==2.3.4
